AWS/utils/sagemaker_fast_api_demo/requirements.txt
pillow
sagemaker
flask
boto3
fastapi
uvicorn